complete dx12 to display editor

+++ Pixel/src/Pixel/Renderer/Framebuffer.h
@@ -65,7 +65,7 @@ namespace Pixel {
 		virtual void ClearAttachment(uint32_t attachmentIndex, int value) = 0;
 
 		virtual uint32_t GetRenderId() = 0;
-		virtual uint32_t GetColorAttachmentRendererID(uint32_t index = 0) const = 0;
+		virtual uint64_t GetColorAttachmentRendererID(uint32_t index = 0) const = 0;
 
 		virtual uint32_t GetDepthAttachmentRendererID() const = 0;
