+++ Pixel/src/Platform/OpenGL/OpenGLFramebuffer.h
@@ -22,14 +22,13 @@ namespace Pixel {
 		virtual void ClearAttachment(uint32_t attachmentIndex, int value);
 
 		//Color Attachment
-		virtual uint32_t GetColorAttachmentRendererID(uint32_t index) const override { return m_ColorAttachments[index]; }
+		virtual uint64_t GetColorAttachmentRendererID(uint32_t index) const override { return m_ColorAttachments[index]; }
 
 		virtual const FramebufferSpecification& GetSpecification() const override { return m_Specification;}
 
 		virtual uint32_t GetDepthAttachmentRendererID() const override { return m_DepthAttachment; }
 
 		virtual void SetDepthAttachmentRendererID(uint32_t rendererID) override;
-
 
 		virtual void CloseColorAttachmentDraw() override;
